requirements.txt
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
boto3>=1.24.0
cffi==1.15.0
click==7.1.2
cryptography==36.0.1
dnspython==2.2.0
email-validator==1.1.3
Flask
Flask-Babel==2.0.0
Flask-BabelEx==0.9.4
Flask-Foundation==2.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-Testing
Flask-WTF==1.0.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
passlib==1.7.4
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
speaklater==1.3
SQLAlchemy==1.3.20
Werkzeug==1.0.1
WTForms==3.0.1
flask-openapi3
bs4
pytest
pytest-playwright
coverage==5.0.3
openapi3-fuzzer
adal==1.2.2
schemathesis